requirements.txt
tensorflow==2.0.1 
tensorflow-hub 
tensorflow_datasets
spacy
gensim
pandas
numpy
matplotlib
seaborn
sklearn
bs4
Unidecode
word2number
contractions
graphviz
pydot
keras-mxnet
hyperas
jupyterlab
validator_collection
nltk
langdetect
tabulate
flask
waitress
